+++ gusto_embedded/lib/gusto_embedded/models/operations/get_v1_companies_company_id_contractor_payment_groups_request.rb
@@ -16,16 +16,16 @@ module GustoEmbedded
       # The time period for which to retrieve contractor payment groups. Defaults to today's date.
       field :end_date, T.nilable(::String), { 'query_param': { 'field_name': 'end_date', 'style': 'form', 'explode': true } }
       # The page that is requested. When unspecified, will load all objects unless endpoint forces pagination.
-      field :page, T.nilable(::Float), { 'query_param': { 'field_name': 'page', 'style': 'form', 'explode': true } }
+      field :page, T.nilable(::Integer), { 'query_param': { 'field_name': 'page', 'style': 'form', 'explode': true } }
       # Number of objects per page. For majority of endpoints will default to 25
-      field :per, T.nilable(::Float), { 'query_param': { 'field_name': 'per', 'style': 'form', 'explode': true } }
+      field :per, T.nilable(::Integer), { 'query_param': { 'field_name': 'per', 'style': 'form', 'explode': true } }
       # The time period for which to retrieve contractor payment groups. Defaults to 6 months ago.
       field :start_date, T.nilable(::String), { 'query_param': { 'field_name': 'start_date', 'style': 'form', 'explode': true } }
       # Determines the date-based API version associated with your API call. If none is provided, your application's [minimum API version](https://docs.gusto.com/embedded-payroll/docs/api-versioning#minimum-api-version) is used.
       field :x_gusto_api_version, T.nilable(::GustoEmbedded::Shared::VersionHeader), { 'header': { 'field_name': 'X-Gusto-API-Version', 'style': 'simple', 'explode': false } }
 
 
-      sig { params(company_id: ::String, end_date: T.nilable(::String), page: T.nilable(::Float), per: T.nilable(::Float), start_date: T.nilable(::String), x_gusto_api_version: T.nilable(::GustoEmbedded::Shared::VersionHeader)).void }
+      sig { params(company_id: ::String, end_date: T.nilable(::String), page: T.nilable(::Integer), per: T.nilable(::Integer), start_date: T.nilable(::String), x_gusto_api_version: T.nilable(::GustoEmbedded::Shared::VersionHeader)).void }
       def initialize(company_id: nil, end_date: nil, page: nil, per: nil, start_date: nil, x_gusto_api_version: nil)
         @company_id = company_id
         @end_date = end_date
